# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Docs api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/offers/": {
            "get": {
                "tags": [
                    "Offers"
                ],
                "summary": "Get product logs",
                "description": "Retrieves product logs based on query parameters.",
                "operationId": "getOffers",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for product logs",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "seller_name",
                        "in": "query",
                        "description": "Seller name to filter by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country to filter by",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Offers"
                ],
                "summary": "Create product logs",
                "description": "Creates new product logs based on provided data.",
                "operationId": "storeOffer",
                "requestBody": {
                    "description": "Product log data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "seller_id",
                                                "amount",
                                                "country"
                                            ],
                                            "properties": {
                                                "seller_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "US"
                                                },
                                                "archive": {
                                                    "type": "string",
                                                    "format": "binary"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data processed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data processed successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductLog"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/offers//{id}": {
            "put": {
                "tags": [
                    "Offers"
                ],
                "summary": "Update a product log",
                "description": "Updates an existing product log.",
                "operationId": "updateProductLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product log to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Product log data to update",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": 150
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "GB"
                                    },
                                    "on_sale": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "archive": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data has been successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data has been\n     *  *             successfully updated"
                                        },
                                        "offer": {
                                            "$ref": "#/components/schemas/ProductLog"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Offers"
                ],
                "summary": "Delete a product log",
                "description": "Deletes an existing product log.",
                "operationId": "deleteProductLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product log to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data has been successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data has been successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/offers//hot": {
            "get": {
                "tags": [
                    "Offers"
                ],
                "summary": "Get hot product logs",
                "description": "Retrieves a list of hot product logs.",
                "operationId": "getHotProductLogs",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductLog"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/offers": {
            "get": {
                "tags": [
                    "Offer Log"
                ],
                "summary": "Retrieve a list of product logs",
                "operationId": "274bc26e0237ae844b20f48cb8a5700b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "seller_name",
                        "in": "query",
                        "description": "Seller name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "post": {
                "tags": [
                    "Offer Log"
                ],
                "summary": "Store a new product log",
                "operationId": "a9fc7c8b6ad8a1a7534fa0b981213b93",
                "requestBody": {
                    "description": "Product log data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "seller_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": "100"
                                                },
                                                "country": {
                                                    "type": "string",
                                                    "example": "USA"
                                                },
                                                "archive": {
                                                    "type": "string",
                                                    "format": "binary"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data processed successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductLog": {
                "required": [
                    "seller_id",
                    "amount",
                    "archive_link",
                    "country"
                ],
                "properties": {
                    "seller_id": {
                        "description": "The ID of the seller associated with the product log",
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "description": "The amount or value associated with the product log",
                        "type": "integer",
                        "example": 100
                    },
                    "archive_link": {
                        "description": "URL to the archive containing product log details",
                        "type": "string",
                        "example": "http://example.com/archives/productlog1.zip"
                    },
                    "is_sold": {
                        "description": "Indicates whether the product log has been sold",
                        "type": "boolean",
                        "example": false
                    },
                    "country": {
                        "description": "The country associated with the product log",
                        "type": "string",
                        "example": "US"
                    },
                    "on_sale": {
                        "description": "Indicates whether the product log is available for sale",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}